Updated the interface names.
