+++ src/main/java/com/welldev/TechBox/authentication/configuration/jwt/JwtService.java
@@ -42,7 +42,7 @@ public class JwtService {
                 .setSubject(userDetails.getUsername())
                 .setIssuedAt(new Date(System.currentTimeMillis()))
                 // token is valid for 10 minutes. It is converted into MiliSeconds.
-                .setExpiration(new Date(System.currentTimeMillis() + 1000 * 60 *60))
+                .setExpiration(new Date(System.currentTimeMillis() + 1000 * 60 * 60))
                 .signWith(getSignInKey(), SignatureAlgorithm.HS256)
                 .compact();
     }
